pyproject.toml

[project]
name = "github-stats-analyser"
dynamic = ["version"]
requires-python = ">=3.13"
dependencies = [
  "gitpython==3.1.44",
  "pygithub==2.6.1",
  "structlog==25.3.0",
  "pygments==2.19.1",
  "pandas==2.2.3",
  "tabulate==0.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.5",
  "pytest-cov==6.1.1",
  "ruff==0.11.9",
  "vulture==2.14",
]
test = [
  "check-jsonschema==0.33.0",
  "requests==2.32.3",
  "Markdown==3.8",
  "beautifulsoup4==4.13.4",
]

[tool.uv]
required-version = "0.7.3"
package = false

[tool.setuptools]
py-modules = ["analyser", "tests"]

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
extend-select = ["E501"]
select = ["ALL"]

ignore = [
  "COM812",  # Ignore due to conflict with Ruff formatter
  "ISC001",  # Ignore due to conflict with Ruff formatter
  "PLR2004", # Ignore magic value
  "D104",    # Ignore missing docstring in public package
  "D100",    # Ignore missing docstring in public module
  "N999",    # Ignore invalid module name
  "SIM112",  # Ignore Lowercase environment variables (used for GitHub actions)
]

fixable = ["ALL"]
unfixable = []

exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
]

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"**test_*.py" = ["S101", "D102", "D103", "SLF001", "PT019"]
"python_scripts/*.py" = ["T201"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["analyser"]

[tool.pytest.ini_options]
norecursedirs = "cloned_repositories"

[tool.vulture]
ignore_names = ["side_effect", "file_paths"]
